Move definition of base_type_eqt to .cpp
This is an implementation detail that need not be exposed to users - the
interface is the free function base_type_eq.
